requirements.txt
dlib
mmcv
scipy
numpy
matplotlib
opencv_python
Pillow
torch>=0.4.0
